Split plugin phase out of executor.go
